Use pv->pv when pv->move_type is SPINDLE
Use the previous-previous move to determine junction velocity, when the
previous move is a SPINDLE speed change (e.g., previous move is not an
actual move). This ensures the unit directions are accurate to the
machine movement, since the SPINDLE move uses "unit" to hold the target
spindle speed, not the actual direction of movement.